Store state.get('documentList')
Before it was state.documentListParams. Setting new params was confusing
because the DocumentListParams could only be built from the documentSet
(which was in DocumentSet) and the view (which was in State). Now that
both are in State, State can build DocumentListParams on its own,
meaning it can build DocumentLists on its own. Yay.