Selection simplification [refs #92354552]
* rename Selection to InMemorySelection
* rename SelectionLike to Selection
* nix unused Selection.timestamp
* nix unused Selection.request
* include documentSetId in Selection Redis keys

The first four are simplifications: Selection is now a persisted Array
of Longs. The last change will increase security when we (soon) let
users specify selections by ID instead of just by query params. Access
to the ID alone (visible in logs) will not give access to the documents:
the user must have access to the document set, too.

It's also perfect for sharding, if we ever shard Redis. Specifying
documentSetId is becoming something of a norm for us.